# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "description": "Retrieves users with filtering by role and pagination. Response detail depends on user role.",
        "operationId": "UsersController_findUsers",
        "parameters": [
          {
            "name": "roleSlug",
            "required": false,
            "in": "query",
            "description": "Filter by user role slug",
            "schema": {
              "enum": [
                "admin",
                "community",
                "supporter",
                "member"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for user name or email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users (detail depends on role)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedUsersResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/PaginatedUsersBasicResponseDto"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (Requires Admin, Member, or Supporter role)"
          },
          "500": {
            "description": "Error retrieving users"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get list of users",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Voluntia API",
    "description": "API documentation for the Voluntia platform backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}